% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amrElm.R
\name{predict.amrElm}
\alias{predict.amrElm}
\title{AMR-ELM.}
\usage{
\method{predict}{amrElm}(model, newdata)
}
\arguments{
\item{model}{the model created by amrElm-train, which contains:
Z: hidden layer weights
H: hidden layer output
weights: output layer weights
affinity: the affinity used to generate the model (e.g.: cosine affinity)
dataTrain: training data for generating affinity matrix. TODO: try to remove X from
the parameter list,
replacing it by the affinity matrix computed for the training patterns}

\item{newdata}{the test data, numerical - numerical with zero mean and unit
variance and patterns in the lines, attributes in the columns}
}
\value{
outupt for the test patterns
}
\description{
Tests the AMR-ELM model.
}
\examples{
\dontrun{testOutput <- predict(model, newdata)}

}
