% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amrElm-test.R
\name{amrElmTest}
\alias{amrElmTest}
\title{Tests the AMR-ELM model.}
\usage{
amrElmTest(XTest, model)
}
\arguments{
\item{XTest}{the test data, numerical - numerical with zero mean and unit 
variance and patterns in the lines, attributes in the columns}

\item{model}{the model created by amrElm-train, which contains:
Z: hidden layer weights 
H: hidden layer output
W: output layer weights 
affinity: the affinity used to generate the model (e.g.: cosine affinity)
dataTrain: training data for generating affinity matrix. TODO: try to remove XTrain from 
the parameter list, 
replacing it by the affinity matrix computed for the training patterns}
}
\value{
outupt for the test patterns
}
\description{
Tests the AMR-ELM model.
}
\examples{
\dontrun{testOutput <- amrElmTest(XTest, model)}
}

