% Generated by roxygen2 (4.0.2): do not edit by hand
\name{amrElmTrain}
\alias{amrElmTrain}
\title{Creates a model for AMR-ELM.}
\usage{
amrElmTrain(l, XTrain, YTrain, affinity = "cosine")
}
\arguments{
\item{l}{the number of hidden neurons}

\item{XTrain}{training data, numerical with zero mean and unit variance and patterns in the lines, attributes in the columns}

\item{YTrain}{training data labels (binary, -1 and +1)}

\item{affinity}{- only cosine implemented}
}
\value{
the amrElm model for supervised problems, with:
         Z: hidden layer weights
         H: hidden layer output
         W: output layer weights
         affinity: the affinity used to generate the model (e.g.: cosine affinity)
         dataTrain: training data for generating affinity matrix.
}
\description{
Creates a model for AMR-ELM.
}
\examples{
data(heart)
data <- heart$data
labels <- heart$labels
l <- 500
N <- nrow(data)
randomPatterns <- seq(N)
data <- data[randomPatterns,]
labels <- labels[randomPatterns]
nTrain <- floor(2*N/3)
nTest <- N - nTrain
data <- data[randomPatterns,]
labels <- labels[randomPatterns]
XTrain <- data[1:nTrain,]
XTest <- data[(nTrain+1):N,]
YTrain <- labels[1:nTrain]
model <- amrElmTrain(l,XTrain,YTrain)
testOutput <- amrElmTest(XTest, model)
}

