% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cosine}
\alias{cosine}
\title{Cosine similarity implementation from http://brainchronicle.blogspot.com.br/2012/06/rcpp-vs-r-implementation-of-cosine.html}
\usage{
cosine(X)
}
\arguments{
\item{X}{numerical data matrix - patterns are in the lines, attributes are in the columns}
}
\value{
Cosine similarity matrix
}
\description{
Cosine similarity implementation from http://brainchronicle.blogspot.com.br/2012/06/rcpp-vs-r-implementation-of-cosine.html
}

